refactor: Move core logic to main script, move map to modules
